SysUI: Use VectorDrawable's for rendering battery
The current implementation of BatteryMeterView uses arrays of points
that are drawn programmatically.  This limits the ability to overlay
or theme the battery.  By utilizing VectorDrawables and
AnimatedVectorDrawables we can achieve the same look and feel of
stock while allowing for more unique battery meters.

Side note:
Dropped dotted circle for the time being
Change-Id: I16a8a949a92827a9e8c1b528981031b6830e97cd